utils: move sendMessageToAdminRoom to utils
